Update AccessCodes.java
Re-organized the code. Put all the overloaded things at the bottom. If we decide to use the overloaded version for the demo with the two Controllers, for the overloaded methods changeMgrCode(String newMgrCode, int accessCodesNumber) and changeUserCode(String newUserCode, int accessCodesNumber), we may need to change, on lines 142 and 173, 
String fileName = "Keypad/resources";
to something else, depending on the structuring of the files we are running our code from.